% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack.R
\name{knapsack}
\alias{knapsack}
\title{Solve 0-1 Knapsack Problem #'
Given the three parameters; wt, vt, w, the function create the matrix. Each cell includes the best value given some restraints}
\usage{
knapsack(wt, vt, w)
}
\arguments{
\item{wt}{represents each item's weight}

\item{vt}{represents each item's value}

\item{w}{represents the weight limit of Knapsack. Total weight of items inside the knapsack cannot exceed this value}
}
\value{
the maximum value in the matrix
}
\description{
Solve 0-1 Knapsack Problem #'
Given the three parameters; wt, vt, w, the function create the matrix. Each cell includes the best value given some restraints
}
